requirements.txt
backports.zoneinfo==0.2.1
greenlet==1.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.36
